reorg constructors
