+++ cloud/cloud_test.go
